long_term/requirements.txt
boto3
pandas
python-dotenv